refactor(db): split db.rs into entity submodules and replace tuple queries (#10)
- Split monolithic db.rs (1277 lines) into db/mod.rs, application.rs,
  interview.rs, task.rs, stage.rs, status.rs, stats.rs
- Replace all manual tuple unpacking with sqlx::FromRow structs
- Merge two-query pattern in get/list_applications into single query
- Eliminate all #[allow(clippy::type_complexity)] in db layer
- Eliminate all #[allow(clippy::too_many_lines)] in db layer
- Eliminate all #[allow(clippy::cast_sign_loss/cast_possible_truncation)]

Closes #10
